refactor(schema): scaffold schemes/lifecycle/ + register stub LifecycleScheme [LTV-Pg.1]
First half of the schema reorg (LTV-Pg). Gives the lifecycle scheme its own
home and makes it a registered peer of lead_scoring, ahead of building its
pipeline (M3–M6). Byte-identical; lead-scoring catalog unchanged.

- New leadforge/schemes/lifecycle/ package:
  - entities.py — the 5 lifecycle rows (CustomerLifecycleRow,
    SubscriptionLifecycleRow, SubscriptionEventRow, HealthSignalRow, InvoiceRow)
    + LIFECYCLE_ROW_TYPES / LIFECYCLE_TABLE_NAMES, moved from schema/entities.py.
    AccountRow / EntityRowProtocol / _empty_df are shared and imported from
    leadforge.schema.entities.
  - relationships.py — LIFECYCLE_CONSTRAINTS, moved from schema/relationships.py
    (reuses the shared FKConstraint).
  - __init__.py — stub LifecycleScheme (build_world/write_bundle raise
    NotImplementedError until M3–M6); self-registers. schemes/__init__ imports it.
- schema/entities.py and schema/relationships.py: lifecycle definitions removed;
  breadcrumb comments point to the new home. ALL_ROW_TYPES / ALL_CONSTRAINTS
  unchanged.
- tests/schema/test_lifecycle_entities.py → tests/schemes/lifecycle/test_entities.py
  with updated imports; tests/schemes/test_registry.py gains lifecycle
  registration + stub-raises-NotImplementedError tests.
- CHANGELOG, CLAUDE.md (both layouts), roadmap (Pg split into Pg.1/Pg.2),
  agent-plan updated.

available_schemes() → ("lead_scoring", "lifecycle"). Verified byte-identical
(14/14 files); full suite 1534 passed / 51 skipped; ruff + mypy clean (92 files).

Note: class-level extraction from the shared schema/entities.py can't be a git
rename (multiple classes pulled from a multi-class file); the lifecycle rows
were only added in #104 so the history loss is shallow.
