+++ .agent-plan.md
@@ -45,10 +45,14 @@ and `LTV-Pe` (#108) merged (scheme protocol + render seam). `LTV-Pf` (physical
 move, **hard break / no shims** per D12): Pf.1 (compute core —
 simulation/mechanisms/structure) merged (#109); Pf.2 (lead-scoring render —
 snapshots/relational_snapshot_safe/tasks moved + relational.py split so the
-shared write_relational_tables stays in the envelope) opened as **#110**. Both
-byte-identical. Sibling `leadforge-datasets-private` consumes bundle files, not
-internals — no lockstep update needed (heads-up issue #8 filed). Next:
-`LTV-Pg` (scaffold `schemes/lifecycle/` + relocate the lead-scoring schema specs).
+shared write_relational_tables stays in the envelope) merged (#110). `LTV-Pg`
+(schema reorg) split: Pg.1 (scaffold `schemes/lifecycle/` — lifecycle rows +
+registries + LIFECYCLE_CONSTRAINTS moved there, stub LifecycleScheme registered)
+opened as **#NNN**; Pg.2 (split the lead-scoring schema out of shared `schema/`)
+pending. All M2 moves byte-identical. Sibling `leadforge-datasets-private`
+consumes bundle files, not internals — no lockstep update needed (heads-up
+issue #8). Next: `LTV-Pg.2`, then `LTV-Pc` (pLTV feature/task specs, authored in
+`schemes/lifecycle/`), then `LTV-M3` (lifecycle population).
 
 ---
