+++ leadforge/schemes/__init__.py
@@ -9,6 +9,7 @@ from __future__ import annotations
 
 # Import built-in scheme modules for their registration side effects.
 from leadforge.schemes import lead_scoring as _lead_scoring  # noqa: F401
+from leadforge.schemes import lifecycle as _lifecycle  # noqa: F401
 from leadforge.schemes.base import (
     SCHEME_REGISTRY,
     GenerationScheme,
